+++ tests/system/tests.py
@@ -5,7 +5,7 @@ import subprocess
 import pytest
 
 import containers
-
+from containers import KEA_1_6, KEA_LATEST
 
 SUPPORTED_DISTROS = [
     ('ubuntu/18.04', 'centos/7'),
@@ -80,19 +80,20 @@ def test_pkg_upgrade(distro_agent, distro_server):
         address=agent.mgmt_ip,
         agentPort=8080)
     r = server.api_post('/machines', json=machine, expected_status=200)  # TODO: POST should return 201
-    assert r.json()['address'] == agent.mgmt_ip
+    data = r.json()
+    assert data['address'] == agent.mgmt_ip
+    m_id = data['id']
 
     for i in range(100):
-        r = server.api_get('/machines')
+        r = server.api_get('/machines/%d/state' % m_id)
         data = r.json()
-        if len(data['items']) == 1 and data['items'][0]['apps'] and len(data['items'][0]['apps'][0]['details']['daemons']) > 1:
+        if data['apps'] and len(data['apps'][0]['details']['daemons']) > 1:
             break
         time.sleep(2)
 
-    m = data['items'][0]
-    assert m['apps'] is not None
-    assert len(m['apps']) == 1
-    assert m['apps'][0]['version'] == '1.7.3'
+    assert data['apps'] is not None
+    assert len(data['apps']) == 1
+    assert data['apps'][0]['version'] == KEA_LATEST.split('-')[0]
 
 
 @pytest.mark.parametrize("agent, server", [('ubuntu/18.04', 'centos/7')])
@@ -118,19 +119,19 @@ def test_add_kea_with_many_subnets(agent, server):
         address=agent.mgmt_ip,
         agentPort=8080)
     r = server.api_post('/machines', json=machine, expected_status=200)  # TODO: POST should return 201
-    assert r.json()['address'] == agent.mgmt_ip
+    data = r.json()
+    assert data['address'] == agent.mgmt_ip
+    m_id = data['id']
 
     for i in range(100):
-        r = server.api_get('/machines')
-        data = r.json()
-        if len(data['items']) == 1 and data['items'][0]['apps']:
+        r = server.api_get('/machines/%d/state' % m_id)
+        m = r.json()
+        if m['apps'] and len(m['apps'][0]['details']['daemons']) > 1:
             break
         time.sleep(2)
-    assert len(data['items']) == 1
-    m = data['items'][0]
     assert m['apps'] is not None
     assert len(m['apps']) == 1
-    assert m['apps'][0]['version'] == '1.7.3'
+    assert m['apps'][0]['version'] == KEA_LATEST.split('-')[0]
     assert len(m['apps'][0]['accessPoints']) == 1
     assert m['apps'][0]['accessPoints'][0]['address'] == '127.0.0.1'
 
@@ -222,3 +223,104 @@ def test_change_kea_ca_access_point(agent, server):
     assert len(m['apps']) == 1
     assert len(m['apps'][0]['accessPoints']) == 1
     assert m['apps'][0]['accessPoints'][0]['address'] == agent.mgmt_ip
+
+
+def run_perfdhcp(src_cntr, dest_ip_addr):
+    cmd = '/usr/sbin/perfdhcp -4 -r %d -R %d -p 10 -b mac=%s:00:00:00:00 %s' % (10, 10000, '00:00', dest_ip_addr)
+    result = src_cntr.run(cmd, ignore_error=True)
+    if result[0] not in [0, 3]:
+        raise Exception('perfdhcp erred: %s' % str(result))
+
+
+# TODO: the test is disabled for now because it does not work on GitLab CI because whole network inside LXD
+# is IPv6 but it is needed to be IPv4
+@pytest.mark.parametrize("agent_kea, agent_old_kea, server", [('ubuntu/20.04', 'ubuntu/18.04', 'centos/7')])
+def atest_get_kea_stats(agent_kea, agent_old_kea, server):
+    elems = [(agent_kea, KEA_LATEST, agent_old_kea.mgmt_ip),
+             (agent_old_kea, KEA_1_6, agent_kea.mgmt_ip)]
+    for idx, (a, ver, relay_addr) in enumerate(elems):
+        a.install_kea(kea_version=ver)
+        a.upload('kea-dhcp4.conf', '/etc/kea/kea-dhcp4.conf')
+        # set proper relay address
+        a.run('sed -i -e s/172\.100\.0\.200/%s/g /etc/kea/kea-dhcp4.conf' % relay_addr)
+        # differentiate subnets which are used in this test
+        prefix = '192.%d.2' % idx
+        a.run('sed -i -e s/192.0.2/%s/g /etc/kea/kea-dhcp4.conf' % prefix)
+        a.run('systemctl restart isc-kea-dhcp4-server')
+
+    r = server.api_post('/sessions', json=dict(useremail='admin', userpassword='admin'), expected_status=200)  # TODO: POST should return 201
+    assert r.json()['login'] == 'admin'
+
+    # add machine: kea and old_kea
+    banner("ADD MACHINES")
+    machines = []
+    for addr in [agent_kea.mgmt_ip, agent_old_kea.mgmt_ip]:
+        machine = dict(
+            address=addr,
+            agentPort=8080)
+        r = server.api_post('/machines', json=machine, expected_status=200)  # TODO: POST should return 201
+        m = r.json()
+        assert m['address'] == addr
+        machines.append(m)
+
+    # wait for discovering apps
+    banner("WAIT FOR DISCOVERING APPS")
+    for i in range(60):
+        # force refreshing machines' state
+        for m in machines:
+            server.api_get('/machines/%d/state' % m['id'])
+        # get machines and check if there is expected data
+        r = server.api_get('/machines')
+        data = r.json()
+        if len(data['items']) == 2 and data['items'][0]['apps'] is not None and data['items'][1]['apps'] is not None:
+            break
+        time.sleep(2)
+
+    # check apps
+    for m in data['items']:
+        assert m['apps'] and len(m['apps']) == 1
+    latest_ver = KEA_LATEST.split('-')[0]
+    assert ((data['items'][0]['apps'][0]['version'] == '1.6.3' and data['items'][1]['apps'][0]['version'] == latest_ver) or
+            (data['items'][0]['apps'][0]['version'] == latest_ver and data['items'][1]['apps'][0]['version'] == '1.6.3'))
+
+    # send DHCP traffic to Kea apps
+    banner("SEND DHCP TRAFFIC TO KEA APPS")
+
+    # send DHCP traffic to old kea
+    agent_kea.run('systemctl stop isc-kea-dhcp4-server')
+    run_perfdhcp(agent_kea, agent_old_kea.mgmt_ip)
+    agent_kea.run('systemctl start isc-kea-dhcp4-server')
+
+    # send DHCP traffic to new kea
+    agent_old_kea.run('systemctl stop isc-kea-dhcp4-server')
+    run_perfdhcp(agent_old_kea, agent_kea.mgmt_ip)
+    agent_old_kea.run('systemctl start isc-kea-dhcp4-server')
+
+    # check gathered stats by Stork server
+    for i in range(80):
+        r = server.api_get('/overview')
+        data = r.json()
+        if data['dhcp4Stats'] and 'assignedAddresses' in data['dhcp4Stats'] and data['dhcp4Stats']['assignedAddresses'] > 150:
+            break
+        time.sleep(2)
+
+    assert data['dhcp4Stats']
+    assert 'assignedAddresses' in data['dhcp4Stats']
+    assert data['dhcp4Stats']['assignedAddresses'] > 150
+    assert 'subnets4' in data
+    assert data['subnets4']['items']
+
+    # there should be 2 subnets where are assigned addresses
+    sn_with_addrs = 0
+    for sn in data['subnets4']['items']:
+        print('=== SN %s: %s' % (sn['subnet'], str(sn)))
+        if sn['subnet'] not in ['192.0.2.0/24', '192.1.2.0/24']:
+            continue
+        for lsn in sn['localSubnets']:
+            print('--- LSN %s' % str(lsn))
+            if 'stats' not in lsn:
+                continue
+            stats = lsn['stats']
+            if 'assigned-addresses' in stats and stats['assigned-addresses'] > 70:
+                sn_with_addrs += 1
+    assert sn_with_addrs == 2